evidently_grafana_project/requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Drift monitoring
evidently==0.4.10

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-docker==3.7.3

# Model persistence
joblib==1.3.2
pickle-mixin==1.0.2

# Monitoring and logging
prometheus-client==0.17.1
python-json-logger==2.0.7

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# API
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
